% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModell.R
\name{logitMod}
\alias{logitMod}
\title{Interface for an alternative logistic regression implementation}
\usage{
logitMod(formula, data)
}
\arguments{
\item{formula}{a formula object. Matrix or array are not accepted}

\item{data}{a data frame contains all variables}
}
\value{
a list of class "logitMod" containing the coefficients, a null model, formula, call,
the dependent variable and all independent variables.
}
\description{
This function computes coefficients of a binary logistic regression
and contructs an object of "logitMod" class
}
\examples{
testData <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")[1:100,]
testData$rank <- factor(testData$rank)
testModell <- as.formula("admit ~ gre + gpa + rank")
testModelFrame <- model.frame(admit ~ gre + gpa + rank, testData)
logitMod(formula = admit ~ gre + gpa + rank, data = testData)
}
