% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModell.R
\name{maxLikeEst}
\alias{maxLikeEst}
\title{Maximum Likelihood}
\usage{
maxLikeEst(y, X)
}
\arguments{
\item{y}{a matrix/vector containing the dependent variable}

\item{X}{a matrix containing the intercept and all independent variables}
}
\value{
a list with maximum likelihood estimation results
}
\description{
This function calculates the maximum likelihood for binary logistic regression
}
\examples{
 
testData <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")[1:100,]
testData$rank <- factor(testData$rank)
testModell <- as.formula("admit ~ gre + gpa + rank")
testModelFrame <- model.frame(admit ~ gre + gpa + rank, testData)
maxLikeEst(X = model.matrix(testModell, testModelFrame), y = model.response(testModelFrame))
}
