% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModell.R
\name{print.summary.logitMod}
\alias{print.summary.logitMod}
\title{Printing method for the summary of "logitMod" estimations}
\usage{
\method{print}{summary.logitMod}(x, ...)
}
\arguments{
\item{x}{an object of class "logitMod"}

\item{...}{unused parameter, methods are required to have same
arguments as their generic functions}
}
\value{
an equivalent output to the summary of the built in binary logistic regression
}
\description{
Printing method for summary of class "logitMod"
}
\examples{
testData <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")[1:100,]
testData$rank <- factor(testData$rank)
testModell <- as.formula("admit ~ gre + gpa + rank")
testModelFrame <- model.frame(admit ~ gre + gpa + rank, testData)
logm <- logitMod(formula = admit ~ gre + gpa + rank, data = testData)
summary(logm)
}
