% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModell.R
\name{summary.logitMod}
\alias{summary.logitMod}
\title{Summary method for "logitMod" estimations}
\usage{
\method{summary}{logitMod}(object, ...)
}
\arguments{
\item{object}{an object of class "logitMod"}

\item{...}{unused parameter, methods are required to have same
arguments as their generic functions}
}
\value{
a list of all necessary values equivalent to the summary output of the
built in binary logistic regression
}
\description{
Summary method for class "logitMod"
}
\examples{
testData <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")[1:100,]
testData$rank <- factor(testData$rank)
testModell <- as.formula("admit ~ gre + gpa + rank")
testModelFrame <- model.frame(admit ~ gre + gpa + rank, testData)
logm <- logitMod(formula = admit ~ gre + gpa + rank, data = testData)
summary(logm)
}
